Added event report calls

+++ src/main/java/seatsio/charts/Category.java
@@ -4,11 +4,11 @@ import seatsio.util.ValueObject;
 
 public class Category extends ValueObject {
 
-    private final int key;
+    private final long key;
     private final String label;
     private final String color;
 
-    public Category(int key, String label, String color) {
+    public Category(long key, String label, String color) {
         this.key = key;
         this.label = label;
         this.color = color;
